requirements.txt
fastapi
uvicorn
sqlmodel
asyncpg
python-multipart
python-dotenv
passlib[bcrypt]
PyJWT
